Consolidate client examples and use OpenEnv rollout
