+++ INSTALL_CLIENT.sh
@@ -233,5 +233,5 @@ echo "  conda activate $ENV_NAME"
 echo ""
 echo "Example usage (see cookbook/client/tinker/):"
 echo "  export MODELSCOPE_TOKEN='your-token'"
-echo "  python cookbook/client/tinker/modelscope_service/self_cognition.py"
+echo "  python cookbook/client/tinker/self_cognition.py"
 echo ""
