+++ cookbook/client/tinker/dpo.py
@@ -39,9 +39,9 @@ from tinker import ServiceClient  # noqa: E402 (must follow init_tinker_client)
 # ---------------------------------------------------------------------------
 # Configuration
 # ---------------------------------------------------------------------------
-base_model = 'Qwen/Qwen3.5-4B'
-base_url = 'http://localhost:8000'
-api_key = 'EMPTY_API_KEY'
+base_model = os.environ.get('TWINKLE_MODEL_ID', 'Qwen/Qwen3.5-4B')
+base_url = os.environ.get('TWINKLE_SERVER_URL', 'http://localhost:8000')
+api_key = os.environ.get('TWINKLE_SERVER_TOKEN', 'EMPTY_TOKEN')
 dataset_id = 'ms://hjh0119/shareAI-Llama3-DPO-zh-en-emoji'
 
 batch_size = 4
