+++ cookbook/client/tinker/upload_to_hub.py
@@ -18,15 +18,17 @@ import dotenv
 
 dotenv.load_dotenv('.env')
 
+import os
+
 from twinkle import get_logger, init_twinkle_client
 from twinkle_client.model import MultiLoraTransformersModel
 
 logger = get_logger()
 
 # ── Configuration ─────────────────────────────────────────────────────────────
-base_model = 'Qwen/Qwen3.5-4B'
-base_url = 'http://localhost:8000'
-api_key = 'EMPTY_TOKEN'  # token used for model training / server access
+base_model = os.environ.get('TWINKLE_MODEL_ID', 'Qwen/Qwen3.5-4B')
+base_url = os.environ.get('TWINKLE_SERVER_URL', 'http://localhost:8000')
+api_key = os.environ.get('TWINKLE_SERVER_TOKEN', 'EMPTY_TOKEN')
 
 # Checkpoint to upload: the twinkle:// path returned by training_client.save_state(),
 # e.g. 'twinkle://20260301_142318-Qwen_Qwen3-4B-199d2cdb/weights/my-lora-epoch-0'
