+++ docs/source_zh/使用指引/服务端和客户端/概述.md
@@ -65,29 +65,22 @@ cookbook/
 │   │   └── mock/                   # Mock 后端（CPU-only 快速启动）
 │   │       └── server_config.yaml
 ├── twinkle/                        # Twinkle Client 示例
-│   ├── self_host/                  # 自托管 Server
-│   │   ├── dpo.py                  # DPO 训练客户端
-│   │   ├── multi_modal.py          # 多模态训练客户端
-│   │   ├── sample.py               # 推理采样客户端
-│   │   ├── self_congnition.py      # 自我认知训练客户端
-│   │   └── short_math_grpo.py      # GRPO 数学训练客户端
-│   └── modelscope/                 # ModelScope 托管服务
-│       ├── dpo.py
-│       ├── multi_modal.py
-│       └── self_congnition.py
+│   ├── dpo.py                      # DPO 训练客户端
+│   ├── embedding.py                # Embedding 训练客户端
+│   ├── multi_modal.py              # 多模态训练客户端
+│   ├── multi_turn_rollout.py       # 多轮 rollout 客户端
+│   ├── sample.py                   # 推理采样客户端
+│   ├── self_cognition.py           # 自我认知训练客户端
+│   ├── short_math_grpo.py          # GRPO 数学训练客户端
+│   └── upload_to_hub.py            # Checkpoint 上传客户端
 └── tinker/                         # Tinker Client 示例
-    ├── self_host/                  # 自托管 Server
-    │   ├── dpo.py                  # DPO 训练客户端
-    │   ├── lora.py                 # LoRA 训练客户端
-    │   ├── multi_modal.py          # 多模态训练客户端
-    │   ├── sample.py               # 推理采样客户端
-    │   ├── self_cognition.py       # 自我认知训练客户端
-    │   └── short_math_grpo.py      # GRPO 数学训练客户端
-    └── modelscope/                 # ModelScope 托管服务
-        ├── dpo.py
-        ├── sample.py
-        ├── self_cognition.py
-        └── short_math_grpo.py
+    ├── dpo.py                      # DPO 训练客户端
+    ├── lora.py                     # LoRA 训练客户端
+    ├── multi_modal.py              # 多模态训练客户端
+    ├── sample.py                   # 推理采样客户端
+    ├── self_cognition.py           # 自我认知训练客户端
+    ├── short_math_grpo.py          # GRPO 数学训练客户端
+    └── upload_to_hub.py            # Checkpoint 上传客户端
 ```
 
 运行步骤：
@@ -96,9 +89,22 @@ cookbook/
 # 1. 先启动 Server
 twinkle-server launch -c cookbook/client/server/transformer/server_config.yaml
 
-# 2. 在另一个终端运行 Client（以 Tinker Client 为例）
-python cookbook/client/tinker/self_host/self_cognition.py
+# 2. 在另一个终端配置客户端
+export TWINKLE_SERVER_URL=http://localhost:8000
+export TWINKLE_SERVER_TOKEN=EMPTY_TOKEN
+export TWINKLE_MODEL_ID=Qwen/Qwen3.5-4B
+
+# 3. 运行 Tinker Client 示例
+python cookbook/client/tinker/self_cognition.py
 
 # 或使用 Twinkle Client
-python cookbook/client/twinkle/self_host/self_cognition.py
+python cookbook/client/twinkle/self_cognition.py
+```
+
+同一组示例也可用于 ModelScope 托管服务，只需切换环境变量：
+
+```bash
+export TWINKLE_SERVER_URL=https://www.modelscope.cn/twinkle
+export TWINKLE_SERVER_TOKEN="$MODELSCOPE_TOKEN"
+export TWINKLE_MODEL_ID=Qwen/Qwen3.6-27B
 ```
